+++ modules/canvaskit/npm_build/types/index.d.ts
@@ -983,6 +983,35 @@ export interface ParagraphBuilder extends EmbindObject<ParagraphBuilder> {
     build(): Paragraph;
 
     /**
+     * Returns a Paragraph object that can be used to be layout and
+     * paint the text to an Canvas.
+     * @param bidiRegions is an array of unsigned integers that should be
+     * treated as triples (starting index, ending index, direction).
+     * Direction == 0 means left-to-right, direction == 1 is right-to-left.
+     * @param words is an array of word edges (starting or ending). You can
+     * pass 2 elements (0 as a start of the entire text and text.size as the
+     * end). This information only needed for a specific API method getWords.
+     * @param graphemes is an array of indexes in the input text that point
+     * to the start of each grapheme.
+     * @param softLineBreaks is an array of indexes in the input text that
+     * point to the places of possible line breaking if needed. It should
+     * include 0 as the first element.
+     * #param hardLineBreaks is an array of indexes in the input text that
+     * point to the places of mandatory line breaking.
+     */
+    buildWithClientInfo(bidiRegions?: InputBidiRegions | null,
+                        words?: | InputWords | null,
+                        graphemes?: InputGraphemes | null,
+                        softLineBreaks?: | InputLineBreaks | null,
+                        hardLineBreaks?: | InputLineBreaks | null): Paragraph;
+
+    /**
+     * Returns the entire Paragraph text (which is useful in case that text
+     * was produced as a set of addText calls).
+     */
+    getText(): string;
+
+    /**
      * Remove a style from the stack. Useful to apply different styles to chunks
      * of text such as bolding.
      */
@@ -4133,6 +4162,24 @@ export type InputFlattenedRSXFormArray = MallocObj | Float32Array | number[];
  * For example, this is the x, y, z coordinates.
  */
 export type InputVector3 = MallocObj | Vector3 | Float32Array;
+
+/**
+ * CanvasKit APIs accept normal arrays, typed arrays, or Malloc'd memory
+ * for bidi regions. Regions are triples of integers
+ * [startIdx, stopIdx, bidiLevel]
+ * where startIdx is inclusive and stopIdx is exclusive.
+ * Length 3 * n where n is the number of regions.
+ */
+export type InputBidiRegions = MallocObj | Uint32Array | number[];
+
+/**
+ * CanvasKit APIs accept normal arrays, typed arrays, or Malloc'd memory for
+ * words, graphemes or line breaks.
+ */
+export type InputWords = MallocObj | Uint32Array | number[];
+export type InputGraphemes = MallocObj | Uint32Array | number[];
+export type InputLineBreaks = MallocObj | Uint32Array | number[];
+
 /**
  * These are the types that webGL's texImage2D supports as a way to get data from as a texture.
  * Not listed, but also supported are https://developer.mozilla.org/en-US/docs/Web/API/VideoFrame
@@ -4172,6 +4219,7 @@ export type RectWidthStyle = EmbindEnumEntity;
 export type TextAlign = EmbindEnumEntity;
 export type TextBaseline = EmbindEnumEntity;
 export type TextDirection = EmbindEnumEntity;
+export type LineBreakType = EmbindEnumEntity;
 export type TextHeightBehavior = EmbindEnumEntity;
 
 export interface AffinityEnumValues extends EmbindEnum {
@@ -4429,6 +4477,11 @@ export interface TextBaselineEnumValues extends EmbindEnum {
 export interface TextDirectionEnumValues extends EmbindEnum {
     LTR: TextDirection;
     RTL: TextDirection;
+}
+
+export interface LineBreakTypeEnumValues extends EmbindEnum {
+    SoftLineBreak: LineBreakType;
+    HardtLineBreak: LineBreakType;
 }
 
 export interface TextHeightBehaviorEnumValues extends EmbindEnum {